requirements/base.txt
Django==4.0.4
djangorestframework==3.13.1
django-environ==0.8.1
django-filter==21.1
django-storages==1.12.3
django-cors-headers==3.11.0
django-extensions==3.1.5
drf-spectacular==0.22.0
kernel-catalogo-videos==0.1.8
python-dateutil==2.8.2
pytz==2022.1
psycopg2==2.9.3
Pillow==9.1.0
google-api-python-client==2.44.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
pydantic==1.9.0
